Products Stock feature in progress, room table has been made, sale feature is using stock use cases

+++ demosales-data/products/src/main/java/com/imecatro/products/data/model/ProductRoomEntity.kt
@@ -11,7 +11,7 @@ data class ProductRoomEntity(
     val price: Float,
     val currency: String,
     val unit: String,
-    val stock: Float,
+    val stock: Double,
     val details: String,
     val imageUri: String
